// frontend/src/api/model/exerciseRead.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Gym Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { ExerciseType } from './exerciseType';

export interface ExerciseRead {
  id: number;
  name: string;
  type: ExerciseType;
  muscle_group_id: number;
  is_active: boolean;
  created_at: string;
}